Cleaned up Makefile.
Declared "all" and "clean" as PHONY.
Made a separate target for the binary, to avoid linking again if everything is up to date.
Let "ld" do the stripping instead of "objcopy".
Do not explicitly add the toolchain include dir, since the cross compiler already looks there.
No need to use "rm -r" if the arguments are never directories.
No need to suppress errors from "rm" if "-f" is given.