Wait until the power slider is released before powering down
Otherwise the unit will reboot instead of powering off.

To notify the user that the unit is powering down, the screen is
disabled, before even "shutdown" is called. That way, the user
will likely release the power slider, and let the unit powers down
for real.

This trick works but is not really smart, and prevent future
development like a "goodbye" splash screen.
I'm looking for better ideas to prevent reboot on shutdown.