Made the brightness module more reliable
The original code leaked file pointers if the key release event was not
received. Also, it lacked error detection in a lot of places and it
contained a lot of duplicated code.

The new code won't open the brightness file again if it is still open.
Also it aborts operations on failure, instead of continuing with a
possibly invalid value. And duplication is reduced by moving the
duplicate code into static functions.

Also simplified the code by using fscanf and fprintf instead of lower
level I/O functions.