Also verify that the event type matches
Previously, a bug was present: the events coming from the analog
stick were considered as part of a combo. So when power was raised
and the analog stick was moved, the poweroff timeout was cancelled
and the BUTTON_POWER key was not sent even though no combo was
actually activated.